Group visitor testing in one file
